extract EntityDirectory component
